Move all routes to /api folder
